# docs/swagger.yaml
basePath: /api/v1/
definitions:
  common.Response:
    properties:
      data_set: {}
      message:
        type: string
      ret_code:
        type: integer
    type: object
  models.Demo:
    properties:
      created_time:
        description: 创建时间
        type: string
      id:
        description: 自增主键
        type: integer
      updated_time:
        description: 更新时间
        type: string
      user_name:
        description: 用户名
        type: string
    type: object
  srvdemo.AddParams:
    properties:
      user_name:
        description: 名字
        type: string
    required:
    - user_name
    type: object
info:
  contact: {}
  description: gin demo
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Demo app
  version: "1.0"
paths:
  /demo/test:
    get:
      consumes:
      - application/json
      description: 获取所有数据
      parameters:
      - description: 自定义查询语句, 使用 RSQL 语法
        in: query
        name: q
        type: string
      - description: 分页条数
        in: query
        name: pagelimit
        type: integer
      - description: 分页偏移量
        in: query
        name: pageoffset
        type: integer
      - description: 关键字模糊查询
        in: query
        name: keyword
        type: string
      - description: 排序, 支持desc和asc, 如 id desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取所有数据
      tags:
      - Demo
    post:
      consumes:
      - application/json
      description: 新增数据
      parameters:
      - description: demo
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/srvdemo.AddParams'
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 新增数据
      tags:
      - Demo
  /demo/test/{id}:
    get:
      consumes:
      - application/json
      description: 获取详情
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data_set:
                  $ref: '#/definitions/models.Demo'
              type: object
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取指定ID详情
      tags:
      - Demo
    put:
      consumes:
      - application/json
      description: 更新数据
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: IDC detail
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/srvdemo.AddParams'
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 更新数据
      tags:
      - Demo
  /demo/test/{ids}:
    delete:
      consumes:
      - application/json
      description: 删除数据
      parameters:
      - description: ids
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除数据
      tags:
      - Demo
  /demo/test/message:
    post:
      consumes:
      - application/json
      description: 发送消息
      parameters:
      - description: demo
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/srvdemo.AddParams'
      produces:
      - application/json
      responses:
        "200":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 结果：{ret_code:code,data:数据,message:消息}
          schema:
            $ref: '#/definitions/common.Response'
      summary: 发送消息
      tags:
      - Demo
swagger: "2.0"